refactor: extract socket wrapping to its own files
