+++ src/client/client.debug.ts
@@ -1,0 +1,4 @@
+
+import Debug from 'debug';
+
+export const debug = Debug('iggy:client');
