+++ src/client/index.ts
@@ -3,3 +3,4 @@ export { Client, SimpleClient, SingleClient } from './client.js'
 export { TcpClient, TcpOption } from './tcp.client.js';
 export { TlsClient, TlsOption } from './tls.client.js';
 export * from './client.utils.js';
+export * from './client.socket.js';
